VaultDetailView refactored
