+++ VultisigApp/VultisigApp/macOS/View/VaultDetailView+macOS.swift
@@ -1,0 +1,67 @@
+//
+//  VaultDetailView+macOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-19.
+//
+
+#if os(macOS)
+import SwiftUI
+
+extension VaultDetailView {
+    var view: some View {
+        list
+            .opacity(showVaultsList ? 0 : 1)
+            .navigationDestination(isPresented: $shouldJoinKeygen) {
+                JoinKeygenView(vault: Vault(name: "Main Vault"))
+            }
+            .navigationDestination(isPresented: $shouldKeysignTransaction) {
+                if let vault = homeViewModel.selectedVault {
+                    JoinKeysignView(vault: vault)
+                }
+            }
+            .navigationDestination(isPresented: $shouldSendCrypto) {
+                SendCryptoView(
+                    tx: sendTx,
+                    vault: vault,
+                    selectedChain: selectedChain
+                )
+            }
+    }
+    
+    var cells: some View {
+        let sortedGroups = viewModel.coinsGroupedByChains.sorted(by: {
+            $0.totalBalanceInFiatDecimal > $1.totalBalanceInFiatDecimal
+        })
+        
+        return ForEach(sortedGroups, id: \.id) { group in
+            ChainNavigationCell(
+                group: group,
+                vault: vault,
+                showAlert: $showAlert
+            )
+        }
+        .background(Color.backgroundBlue)
+        .padding(.horizontal, 16)
+    }
+    
+    var chooseChainButton: some View {
+        NavigationLink {
+            ChainSelectionView(showChainSelectionSheet: $showSheet, vault: vault)
+        } label: {
+            chooseChainButtonLabel
+        }
+        .padding(.horizontal, 16)
+        .frame(height: 20)
+        .font(.body16MenloBold)
+        .foregroundColor(.turquoise600)
+    }
+    
+    var scanButton: some View {
+        VaultDetailScanButton(showSheet: $showScanner, sendTx: sendTx)
+            .opacity(showVaultsList ? 0 : 1)
+            .buttonStyle(BorderlessButtonStyle())
+            .padding(.bottom, 30)
+    }
+}
+#endif
